+++ src/corax/kernel/generic/core_derivatives.c
@@ -713,7 +713,7 @@ corax_core_likelihood_derivatives(unsigned int        states,
                                   unsigned int        parent_sites,
                                   unsigned int        child_ids,
                                   const int *         invariant,
-                                  const unsigned int *pattern_weights,
+                                  const double *      pattern_weights,
                                   double              branch_length,
                                   const double *      prop_invar,
                                   double *const *     freqs,
@@ -862,7 +862,7 @@ corax_core_likelihood_derivatives(unsigned int        states,
   if (attrib & CORAX_ATTRIB_AB_MASK)
   {
     double       asc_Lk[3] = {0.0, 0.0, 0.0};
-    unsigned int sum_w_inv = 0;
+    double       sum_w_inv = 0.0;
     double       asc_scaling;
     int          asc_bias_type = attrib & CORAX_ATTRIB_AB_MASK;
 
@@ -907,7 +907,7 @@ corax_core_likelihood_derivatives(unsigned int        states,
       case CORAX_ATTRIB_AB_LEWIS:
       {
         // TODO: pattern_weight_sum should be stored somewhere!
-        unsigned int pattern_weight_sum = 0;
+        double pattern_weight_sum = 0.0;
         for (n = 0; n < ef_sites; ++n) pattern_weight_sum += pattern_weights[n];
 
         /* derivatives of log(1.0 - (sum Li(s) over states 's')) */
